+++ HISTORY.md
@@ -1,3 +1,18 @@
+1.1.0 / 2022-01-31
+===================
+
+  * Added Javascript API to handle icons import
+
+1.0.4 / 2022-01-09
+===================
+
+  * Fixed handle error http to get icons
+
+1.0.3 / 2022-01-09
+===================
+
+  * Update documentation
+
 1.0.2 / 2022-01-09
 ===================